+++ packages/gatsby-plugin-graphql-codegen/src/gatsby-node.ts
@@ -1,5 +1,9 @@
-import { GatsbyNode, PluginOptions } from 'gatsby'
-import { generateWithConfig } from './graphql-codegen.config'
+import { GatsbyNode } from 'gatsby'
+import {
+  generateWithConfig,
+  GatsbyCodegenPlugins,
+  CodegenOptions,
+} from './graphql-codegen.config'
 import debounce from 'lodash.debounce'
 import { GraphQLTagPluckOptions } from '@graphql-toolkit/graphql-tag-pluck'
 import { loadSchema, UnnormalizedTypeDefPointer } from '@graphql-toolkit/core'
@@ -16,19 +20,19 @@ export interface SchemaConfig {
   schema: UnnormalizedTypeDefPointer
   documentPaths?: string[]
   pluckConfig: GraphQLTagPluckOptions
+  codegenPlugins?: GatsbyCodegenPlugins[]
+  codegenConfig?: Record<string, any>
 }
 
-export interface TsCodegenOptions extends PluginOptions {
-  documentPaths?: string[]
-  fileName?: string
+export interface PluginOptions extends Partial<CodegenOptions> {
+  plugins: unknown[]
   codegen?: boolean
   codegenDelay?: number
-  pluckConfig?: GraphQLTagPluckOptions
   failOnError?: boolean
   additionalSchemas?: SchemaConfig[]
 }
 
-const defaultOptions: Required<TsCodegenOptions> = {
+const defaultOptions: Required<PluginOptions> = {
   plugins: [],
   documentPaths: ['./src/**/*.{ts,tsx}', './node_modules/gatsby-*/**/*.js'],
   fileName: 'graphql-types.ts',
@@ -45,9 +49,11 @@ const defaultOptions: Required<TsCodegenOptions> = {
     ],
   },
   additionalSchemas: [],
+  codegenPlugins: [],
+  codegenConfig: {},
 }
 
-type GetOptions = (options: TsCodegenOptions) => Required<TsCodegenOptions>
+type GetOptions = (options: PluginOptions) => Required<PluginOptions>
 const getOptions: GetOptions = pluginOptions => ({
   ...defaultOptions,
   ...pluginOptions,
@@ -60,9 +66,9 @@ type AsyncMap = <T, TResult>(
 const asyncMap: AsyncMap = (collection, callback) =>
   Promise.all(collection.map(callback))
 
-export const onPostBootstrap: NonNullable<GatsbyNode['onPostBootstrap']> = async (
+export const onPostBootstrap: GatsbyNode['onPostBootstrap'] = async (
   { store, reporter },
-  pluginOptions: TsCodegenOptions
+  pluginOptions: PluginOptions = { plugins: [] }
 ) => {
   const options = getOptions(pluginOptions)
   if (!options.codegen) return
@@ -74,6 +80,8 @@ export const onPostBootstrap: NonNullable<GatsbyNode['onPostBootstrap']> = async
     pluckConfig,
     additionalSchemas,
     failOnError,
+    codegenPlugins,
+    codegenConfig,
   } = options
 
   const {
@@ -90,6 +98,8 @@ export const onPostBootstrap: NonNullable<GatsbyNode['onPostBootstrap']> = async
     directory,
     schema,
     reporter,
+    codegenPlugins,
+    codegenConfig,
   }
 
   const configs = [
@@ -109,6 +119,8 @@ export const onPostBootstrap: NonNullable<GatsbyNode['onPostBootstrap']> = async
             new GraphQLFileLoader(),
           ],
         }),
+        codegenPlugins: [],
+        codegenConfig: {},
         reporter,
         ...config,
       }