+++ packages/gatsby-plugin-graphql-codegen/src/graphql-codegen.config.ts
@@ -1,11 +1,16 @@
 import * as fs from 'fs-extra'
 import * as path from 'path'
 import { Reporter } from 'gatsby'
+import {
+  Types,
+  CodegenPlugin,
+  PluginFunction,
+} from '@graphql-codegen/plugin-helpers'
 import { Source } from '@graphql-toolkit/common'
 import { loadDocuments } from '@graphql-toolkit/core'
 import { CodeFileLoader } from '@graphql-toolkit/code-file-loader'
 import { codegen } from '@graphql-codegen/core'
-import { printSchema, parse } from 'gatsby/graphql'
+import { printSchema, parse, GraphQLSchema } from 'gatsby/graphql'
 import { plugin as typescriptPlugin } from '@graphql-codegen/typescript'
 import { plugin as operationsPlugin } from '@graphql-codegen/typescript-operations'
 import { GraphQLTagPluckOptions } from '@graphql-toolkit/graphql-tag-pluck'
@@ -14,29 +19,115 @@ function isSource(result: void | Source[]): result is Source[] {
   return typeof result !== 'undefined'
 }
 
-interface IInitialConfig {
-  documentPaths: string[]
-  directory: string
-  fileName: string
-  reporter: Reporter
-  pluckConfig: GraphQLTagPluckOptions
+export type GatsbyCodegenPlugins = {
+  resolve: 'typescript' | 'operations' | PluginFunction
+  addToSchema: CodegenPlugin['addToSchema']
+  validate: CodegenPlugin['validate']
+  options: Record<string, any>
 }
 
-type CreateConfigFromSchema = (schema: any) => Promise<any>
-type CreateConfig = (args: IInitialConfig) => Promise<CreateConfigFromSchema>
+type Plugins = {
+  pluginMap: Record<string, CodegenPlugin>
+  plugins: Record<string, Record<string, any>>[]
+}
+
+const defaultPlugins: Plugins = {
+  plugins: [
+    {
+      typescript: {
+        skipTypename: true,
+        enumsAsTypes: true,
+      },
+    },
+    {
+      operations: {
+        skipTypename: true,
+        exportFragmentSpreadSubTypes: true,
+      },
+    },
+  ],
+  pluginMap: {
+    typescript: {
+      plugin: typescriptPlugin,
+    },
+    operations: {
+      plugin: operationsPlugin,
+    },
+  },
+}
+export const mapCodegenPlugins = ({
+  codegenPlugins,
+  defaultPlugins,
+}: {
+  codegenPlugins: GatsbyCodegenPlugins[]
+  defaultPlugins: Plugins
+}): Plugins =>
+  codegenPlugins.reduce((acc, plugin, i) => {
+    const { resolve, options, ...otherOptions } = plugin
+    // handle default plugins (typescript, operations)
+    if (typeof resolve === 'string') {
+      const added = acc.plugins.find(
+        addedPlugin => Object.keys(addedPlugin)[0] === resolve
+      )
+
+      if (!added) {
+        throw new Error(
+          `[gatsby-plugin-graphql-codegen] Invalid codegenPlugins: ${resolve}`
+        )
+      }
+
+      added[resolve] = {
+        ...added[resolve],
+        ...options,
+      }
+      // presumably new plugins
+    } else {
+      const identifier = `codegen-plugin-${i}`
+      acc.plugins.push({ [identifier]: options })
+      acc.pluginMap[identifier] = { plugin: resolve, ...otherOptions }
+    }
+    return acc
+  }, defaultPlugins)
+
+export interface CodegenOptions {
+  documentPaths: string[]
+  fileName: string
+  pluckConfig: GraphQLTagPluckOptions
+  codegenPlugins: GatsbyCodegenPlugins[]
+  codegenConfig: Record<string, any>
+}
+
+interface CreateConfigOptions extends CodegenOptions {
+  directory: string
+  reporter: Reporter
+}
+
+type CreateConfigFromSchema = (
+  schema: GraphQLSchema
+) => Promise<Types.GenerateOptions>
+type CreateConfig = (
+  args: CreateConfigOptions
+) => Promise<CreateConfigFromSchema>
 const createConfig: CreateConfig = async ({
   documentPaths,
   directory,
   fileName,
   reporter,
   pluckConfig,
+  codegenPlugins,
+  codegenConfig,
 }) => {
   // file name & location
   const pathToFile = path.join(directory, fileName)
   const { dir } = path.parse(pathToFile)
   await fs.ensureDir(dir)
 
-  return async (schema): Promise<any> => {
+  const { pluginMap, plugins } = mapCodegenPlugins({
+    codegenPlugins,
+    defaultPlugins,
+  })
+
+  return async (schema): Promise<Types.GenerateOptions> => {
     // documents
     const docPromises = documentPaths.map(async docGlob => {
       const _docGlob = path.join(directory, docGlob)
@@ -55,36 +146,17 @@ const createConfig: CreateConfig = async ({
     return {
       filename: pathToFile,
       schema: parse(printSchema(schema)),
-      plugins: [
-        {
-          typescript: {
-            skipTypename: true,
-            enumsAsTypes: true,
-          },
-        },
-        {
-          typescriptOperation: {
-            skipTypename: true,
-            exportFragmentSpreadSubTypes: true,
-          },
-        },
-      ],
+      config: codegenConfig,
       documents,
-      pluginMap: {
-        typescript: {
-          plugin: typescriptPlugin,
-        },
-        typescriptOperation: {
-          plugin: operationsPlugin,
-        },
-      },
+      plugins,
+      pluginMap,
     }
   }
 }
 
-type GenerateFromSchema = (schema: any) => Promise<void>
+type GenerateFromSchema = (schema: GraphQLSchema) => Promise<void>
 type GenerateWithConfig = (
-  initialOptions: IInitialConfig
+  initialOptions: CreateConfigOptions
 ) => Promise<GenerateFromSchema>
 export const generateWithConfig: GenerateWithConfig = async initialOptions => {
   const createConfigFromSchema = await createConfig(initialOptions)